detect on-iphone compilation and adjust args accordingly, fix ldid to use $BIN
If compiling on the iPhone platform, the -arch option causes gcc to fail.
This commit checks the type of the compiling machine using uname -m, and removes the -arch argument if it finds the string "iphone" in the result of
said check.
Also fix the sign command to use the $BIN path when calling ldid.